add decode header impl

+++ sdr_decode/cpp/types.h
@@ -3,8 +3,8 @@
 #include <valarray>
 
 // https://en.cppreference.com/w/cpp/numeric/complex
-//typedef std::complex<float> complex_t;
-typedef std::complex<double> complex_t;
+typedef std::complex<float> complex_t;
+//typedef std::complex<double> complex_t;
 
 // https://en.cppreference.com/w/cpp/numeric/valarray
 typedef std::valarray<complex_t> cpvarray_t;
